requirements.txt
jsonpath=0.82
paho-mqtt=1.5.0
websockets=8.1